pcap_parsing_examples/requirements.txt
scapy-python3
netifaces
